feat_fix: Se hacen cambios y refactoring sobre la funcionalidad del sistema hasta el momento
