+++ .idea/runConfigurations/Unnamed.xml
@@ -13,7 +13,7 @@
     <option name="port" value="8081" />
     <option name="tomcatInfo">
       <TomcatInfo>
-        <option name="name" value="Apache Tomcat/10.0.23 (3)" />
+        <option name="name" value="Apache Tomcat/10.0.23" />
         <option name="path" value="$PROJECT_DIR$/../../../../../../apache-tomcat-10.0.23" />
         <option name="version" value="10.0.23.0" />
       </TomcatInfo>
