+++ src/main/java/taller2/servlets/HomeServlet.java
@@ -1,0 +1,70 @@
+package taller2.servlets;
+
+import jakarta.servlet.RequestDispatcher;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
+import main.java.taller1.Logica.Clases.Usuario;
+import main.java.taller1.Logica.Fabrica;
+
+import java.io.IOException;
+
+@WebServlet(name = "Home", value = "/")
+public class HomeServlet extends HttpServlet {
+    
+    Fabrica fabrica;
+    
+    public void init() {
+        fabrica = Fabrica.getInstance();
+    }
+    
+    protected void dispatchPage(String page, HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        response.setContentType("text/html;charset=UTF-8");
+        RequestDispatcher view = request.getRequestDispatcher(page);
+        view.forward(request, response);
+    }
+    
+    protected boolean checkSession(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        HttpSession session = request.getSession(false);
+        // Si no hay sesión, redirigir a login
+        if (session == null) {
+            return false;
+        }
+        
+        // Si hay sesión, obtener el usuario
+        Usuario usuarioLogueado = (Usuario) session.getAttribute("usuarioLogueado");
+        
+        // Si no hay usuario, redirigir a login
+        if (usuarioLogueado == null) {
+            return false;
+        }
+        
+        // Si hay usuario, enviarlo a la página de inicio
+        return true;
+    }
+    
+    @Override
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        // Si no hay sesión, redirigir a login
+        boolean sessionIniciada = checkSession(request, response);
+        HttpSession session = request.getSession();
+        if(sessionIniciada) {
+            dispatchPage("/pages/index.jsp", request, response);
+        } else {
+            response.sendRedirect("login");
+        }
+    }
+    
+    @Override
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        // CERRAR SESION
+        HttpSession session = request.getSession();
+        session.invalidate();
+        
+        response.sendRedirect("login");
+    }
+    
+}
