+++ src/main/java/taller2/servlets/Paquete/ListadoPaquetesServlet.java
@@ -1,0 +1,38 @@
+package taller2.servlets.Paquete;
+
+import jakarta.servlet.RequestDispatcher;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import main.java.taller1.Logica.Clases.Paquete;
+import main.java.taller1.Logica.Fabrica;
+
+import java.io.IOException;
+import java.util.Map;
+
+@WebServlet(name = "ListadoPaquetes", value = "/listado-paquetes")
+public class ListadoPaquetesServlet extends HttpServlet {
+    Fabrica fabrica;
+
+    public void init() {
+        fabrica = Fabrica.getInstance();
+    }
+    protected void dispatchPage(String page, HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        response.setContentType("text/html;charset=UTF-8");
+        RequestDispatcher view = request.getRequestDispatcher(page);
+        view.forward(request, response);
+    }
+
+    @Override
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        Map<String, Paquete> paquetes = fabrica.getIPaquete().obtenerPaquetes();
+        request.setAttribute("paquetes", paquetes);
+        dispatchPage("/pages/paquete/listado-paquetes.jsp", request, response);
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+    }
+}
