+++ src/main/java/taller2/servlets/Usuario/ListadoUsuarioServlet.java
@@ -1,0 +1,36 @@
+package taller2.servlets.Usuario;
+
+import jakarta.servlet.RequestDispatcher;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import main.java.taller1.Logica.Clases.Usuario;
+import main.java.taller1.Logica.Fabrica;
+
+import java.io.IOException;
+import java.util.Map;
+
+@WebServlet(name = "ListadoUsuario", value = "/listado-usuario")
+public class ListadoUsuarioServlet extends HttpServlet {
+  Fabrica fabrica;
+  
+  public void init() {
+    fabrica = Fabrica.getInstance();
+  }
+  
+  @Override
+  protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+    response.setContentType("text/html");
+    Map<String, Usuario> usuarios = Fabrica.getInstance().getIUsuario().obtenerUsuarios();
+    request.setAttribute("usuarios", usuarios);
+    RequestDispatcher view = request.getRequestDispatcher("/pages/usuario/listado-usuario.jsp");
+    view.forward(request, response);
+  }
+  
+  @Override
+  protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+  
+  }
+}
