+++ src/main/java/taller2/utils/MainUtils.java
@@ -1,0 +1,14 @@
+package taller2.utils;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class MainUtils {
+
+//  Espacio para poner metodos de utilidad estaticas por si llegamos a necesitar
+  
+  public static void helloWorld() {
+    System.out.println("Hello World!");
+  }
+  
+}
