+++ src/main/webapp/pages/altas.css
@@ -1,0 +1,32 @@
+
+#idform {
+    top: 50%;
+    padding: 30px;
+}
+.subtitulos{
+    width: 30%;
+}
+#idform input {
+    width: 50%;
+    float: left;
+    text-align: left;
+    line-height: 30px;
+}
+#idform textarea {
+    width: 50%;
+    float: left;
+    text-align: left;
+    resize: none;
+    line-height: 60px;
+}
+.input-container{
+    border-bottom: 1px solid #f2f2f2;
+    padding-bottom: 20px;
+    width: 100%;
+    display: flex;
+    -webkit-box-pack: justify;
+    /*justify-content: space-between;*/
+    -webkit-box-align: center;
+    align-items: center;
+    margin-bottom: 15px;
+}
