+++ src/main/webapp/pages/paquete/detalle-paquete.jsp
@@ -2,12 +2,12 @@
 <%@ page import="main.java.taller1.Logica.Clases.*" %>
 <%@ page contentType="text/html; charset=UTF-8" pageEncoding="UTF-8" %>
 <html>
-    <head>
-        <title>Detalle de paquete</title>
-        <style><%@ include file="/pages/global.css" %></style>
-        <style><%@ include file="/pages/detalles.css" %></style>
-    </head>
-    <body>
+<head>
+    <title>Detalle de paquete</title>
+    <style><%@ include file="/pages/global.css" %></style>
+    <style><%@ include file="/pages/detalles.css" %></style>
+</head>
+<body>
     <%@ include file="/pages/header.jsp" %>
     <section>
         <%@ include file="/pages/sidebar.jsp" %>
@@ -17,89 +17,64 @@
             String respuesta = (String) request.getAttribute("respuesta");
         %>
         <div class="grid-container">
-
-
-                <%-- AGREGAR COMPONENTES ACA--%>
-
-                <h1 class="title">Detalle de paquete</h1>
-
-                    <img src="https://cdn-icons-png.flaticon.com/512/44/44248.png" alt="Foto de perfil" class="img_perfil">
+            <%-- AGREGAR COMPONENTES ABAJO--%>
+            <h1 class="title">Detalle de paquete</h1>
+            <img src="https://cdn-icons-png.flaticon.com/512/44/44248.png" alt="Foto de perfil" class="img_perfil">
             <div class="first-data">
                 <h2><%=paquete.getNombre()%></h2>
                 <h4>Fecha de expiracion:<%=paquete.getFechaExpiracion()%></h4>
-
             </div>
-            <%
-                if(respuesta != "Paquete Adquirido"){
-            %>
+            <%  if(respuesta != "Paquete Adquirido"){ %>
                     <form class="form" action="detalle-paquete" method="POST">
                         <input type="hidden" name="nombre" value="<%=paquete.getNombre()%>">
                         <button class="btn2">Comprar Paquete</button>
                     </form>
-            <%
-                }else{
-            %>
+            <%  }else  { %>
                     <h4 class="respuesta"><%=respuesta%></h4>
-            <%
-                }
-            %>
+            <%  } %>
             <div class="tabs">
                 <div class="menu">
-                        <p data-target="#datos_generales" class="active">Datos Generales</p>
-                        <p data-target="#descripcion">Descripcion</p>
-                        <p data-target="#espectaculos">Espectaculos</p>
-
-
+                    <p data-target="#datos_generales" class="active">Datos Generales</p>
+                    <p data-target="#descripcion">Descripcion</p>
+                    <p data-target="#espectaculos">Espectaculos</p>
                 </div>
 
                 <div class="content">
-
                     <div data-content id="datos_generales" class="active">
                         <h4>Nombre:<%=paquete.getNombre()%></h4>
                         <h4>Descuento:<%=paquete.getDescuento()%></h4>
                         <h4>Fecha de registro:<%=paquete.getFechaRegistro()%></h4>
                         <h4>Fecha de expiracion:<%=paquete.getFechaExpiracion()%></h4>
-
                     </div>
 
                     <div data-content id="descripcion">
                         <h4><%=paquete.getDescripcion()%></h4>
-
-
                     </div>
-
-
+                    
                     <div data-content id="espectaculos">
                         <table >
                             <tbody>
-                            <%
-
-                                Map<String, Espectaculo> espectaculos = (Map<String, Espectaculo>) request.getAttribute("espectaculos");
-                                for (Espectaculo espectaculo : espectaculos.values()) {
-                            %>
-                            <tr>
-                                <th onClick="location.href='detalle-espectaculo?nombre=<%=espectaculo.getNombre()%>&plataforma=<%=espectaculo.getPlataforma().getNombre()%>'"> <%=espectaculo.getNombre()%> </th>
-                            </tr>
-                            <%
-                                }
-                            %>
-
+                            <%  Map<String, Espectaculo> espectaculos = (Map<String, Espectaculo>) request.getAttribute("espectaculos");
+                                for (Espectaculo espectaculo : espectaculos.values()) { %>
+                                    <tr>
+                                        <th onClick="location.href='detalle-espectaculo?nombre=<%=espectaculo.getNombre()%>&plataforma=<%=espectaculo.getPlataforma().getNombre()%>'"> <%=espectaculo.getNombre()%> </th>
+                                    </tr>
+                            <%  }   %>
                             </tbody>
                         </table>
                     </div>
-
-
-
                 </div>
             </div>
+    
+            <button class="volver" onclick="history.back()">Volver</button>
         </div>
-
-            <%--                AGREGAR COMPONENTES ACA--%>
-
+        <%-- AGREGAR COMPONENTES ARRIBA--%>
     </section>
 
+    <%--    Javascript--%>
+    <script src="https://code.jquery.com/jquery-3.6.1.min.js" integrity="sha256-o88AwQnZB+VDvE9tvIXrMQaPlFFSUTR+nldQm1LuPXQ=" crossorigin="anonymous"></script>
     <script>
-        var targets = document.querySelectorAll('[data-target]')
+        let targets = document.querySelectorAll('[data-target]')
         const content = document.querySelectorAll('[data-content]')
 
         targets.forEach(target  => {
@@ -119,9 +94,6 @@
                 console.log(target.classList);
             })
         })
-
-
-
     </script>
     </body>
 </html>
