+++ src/main/webapp/pages/usuario/listado-usuario.jsp
@@ -1,7 +1,12 @@
 <%@ page import="main.java.taller1.Logica.Clases.Usuario" %>
 <%@ page import="java.util.Map" %>
 <%@ page import="main.java.taller1.Logica.Clases.Artista" %>
+<%@ page import="com.google.gson.Gson" %>
 <%@ page contentType="text/html; charset=UTF-8" pageEncoding="UTF-8" %>
+<%
+    Map<String, Usuario> usuarios = request.getAttribute("usuarios") != null ? (Map<String, Usuario>) request.getAttribute("usuarios") : null;
+    String json = new Gson().toJson(usuarios);
+%>
 <!DOCTYPE html>
 <html>
 <head>
@@ -11,20 +16,18 @@
     <title>JSP - Hello World</title>
 </head>
 <body>
-<%@ include file="/pages/header.jsp" %>
-
-<section>
-    <%@ include file="/pages/sidebar.jsp" %>
-    <div class="main-container">
-        <%--                AGREGAR COMPONENTES ACA--%>
+    <%@ include file="/pages/header.jsp" %>
+    <section>
+        <%@ include file="/pages/sidebar.jsp" %>
+        <div class="main-container">
+            <%-- AGREGAR COMPONENTES ABAJO--%>
             <div class="busqueda">
-                <br/>
-                <label for="">Buscar Usuario</label>
+                <label for="txtBuscar">Buscar Usuario</label>
                 <input type="text" name="buscarUsuario" id="txtBuscar" value="Nombre...">
             </div>
-            <h2>Usuarios</h2>
-            <br>
             <div>
+                <h2>Usuarios</h2>
+                <button class="volver" onclick="history.back()">Volver</button>
                 <table class="tablaUsuarios" id="tabla">
                     <thead>
                     <tr>
@@ -36,57 +39,69 @@
                     </tbody>
                 </table>
             </div>
-        <%--                AGREGAR COMPONENTES ACA--%>
-    </div>
-</section>
-
-<script src="https://code.jquery.com/jquery-3.6.1.min.js" integrity="sha256-o88AwQnZB+VDvE9tvIXrMQaPlFFSUTR+nldQm1LuPXQ=" crossorigin="anonymous"></script>
-<script>
-    $(document).ready(function(){
-
-    });
-
-    <% Map<String, Usuario> usuarios = (Map<String, Usuario>) request.getAttribute("usuarios");
-    for (Usuario elem : usuarios.values()) {%>
-    nuevaFila = tabla.insertRow(-1);
-    celdaNickname = nuevaFila.insertCell(0);
-    celdaTipo = nuevaFila.insertCell(1);
-
-    celdaNickname.innerHTML = "<%=elem.getNickname()%>";
-    <% if (elem instanceof Artista){%>
-    celdaTipo.innerHTML = "Artista";
-    <%}else{%>
-    celdaTipo.innerHTML = "Espectador";
-    <%}%>
-
-    celdaNickname.setAttribute('onClick',"location.href='detalle-usuario?nickname=<%=elem.getNickname()%>'");
-    <% } %>
-
-    $("#txtBuscar").on("keyup", function() {
-        var keyword = this.value;
-        keyword = keyword.toUpperCase();
-        var table_1 = document.getElementById("tabla");
-        var all_tr = table_1.getElementsByTagName("tr");
-        for(var i=0; i<all_tr.length; i++){
-            var name_column = all_tr[i].getElementsByTagName("td")[0];
-            if(name_column){
-                var name_value = name_column.textContent || name_column.innerText;
-                name_value = name_value.toUpperCase();
-                if(name_value.indexOf(keyword) > -1){
-                    all_tr[i].style.display = ""; // show
-                }else{
-                    all_tr[i].style.display = "none"; // hide
+            <%-- AGREGAR COMPONENTES ARRIBA--%>
+        </div>
+    </section>
+    
+    <%--    Javascript--%>
+    <script src="https://code.jquery.com/jquery-3.6.1.min.js" integrity="sha256-o88AwQnZB+VDvE9tvIXrMQaPlFFSUTR+nldQm1LuPXQ=" crossorigin="anonymous"></script>
+    <script>
+        // Declaramos elementos del DOM
+        const TABLA = document.getElementById("cuerpoTabla");
+        const TXT_BUSCAR = $("#txtBuscar");
+    
+        //CUANDO EL DOCUMENTO ESTE LISTO
+        $(document).ready(function(){
+            crearTabla();
+        });
+    
+    
+        function crearTabla(){
+            let nuevaFila;
+            let celdaNickname;
+            let celdaTipo;
+    
+            <% for (Usuario elem : usuarios.values()) {%>
+                nuevaFila = TABLA.insertRow(-1);
+                celdaNickname = nuevaFila.insertCell(0);
+                celdaTipo = nuevaFila.insertCell(1);
+    
+                celdaNickname.innerHTML = "<%=elem.getNickname()%>";
+                <% if (elem instanceof Artista){%>
+                    celdaTipo.innerHTML = "Artista";
+                <%}else{%>
+                    celdaTipo.innerHTML = "Espectador";
+                <%}%>
+    
+                nuevaFila.addEventListener("click", function(){
+                    window.location.href = "perfil?nickname=<%=elem.getNickname()%>";
+                });
+            <% } %>
+        }
+    
+        TXT_BUSCAR.on("keyup", function() {
+            let keyword = this.value;
+            keyword = keyword.toUpperCase();
+            let table_1 = document.getElementById("tabla");
+            let all_tr = table_1.getElementsByTagName("tr");
+            for(let i=0; i<all_tr.length; i++){
+                let name_column = all_tr[i].getElementsByTagName("td")[0];
+                if(name_column){
+                    let name_value = name_column.textContent || name_column.innerText;
+                    name_value = name_value.toUpperCase();
+                    if(name_value.indexOf(keyword) > -1){
+                        all_tr[i].style.display = ""; // show
+                    }else{
+                        all_tr[i].style.display = "none"; // hide
+                    }
                 }
             }
-        }
-    });
-
-    $("#txtBuscar").click(function(){
-        $("#txtBuscar").val("");
-    });
-
-
-
-</script>
+        });
+    
+        TXT_BUSCAR.click(function(){
+            TXT_BUSCAR.val("");
+        });
+    
+    </script>
 </body>
 </html>
