+++ src/main/webapp/pages/usuario/login.css
@@ -45,8 +45,6 @@
     background-color: rgba(255, 0, 0, 0.5);
 }
 
-
-
 .container {
   display: flex;
   justify-content: center;
@@ -87,11 +85,13 @@
   font-size: 25px;
   font-weight: 600;
   margin-bottom: 10px;
+  text-align: center;
 }
 
 .container__right > h3 > span {
   color: var(--primary-blue);
-  font-weight: bolder
+  font-weight: bolder;
+  font-size: 25px;
 }
 .container__right > h3 > span:hover {
   cursor: pointer;
@@ -111,9 +111,9 @@
   display: flex;
   flex-direction: column;
   justify-content: center;
-  align-items: left;
+  align-items: flex-start;
   width: 80%;
-  height: 80%;
+  margin: 30px 0;
 }
 
 #idform > input {
@@ -144,5 +144,12 @@
 .forgot {
   font-size: 12px;
   color: #00376b;
-  text-align: right;
+  text-align: right !important;
+}
+
+@media screen and (max-height: 610px) {
+  .container {
+    width: 60%;
+    height: 70%;
+  }
 }
