+++ src/main/webapp/pages/usuario/registro.css
@@ -1,0 +1,157 @@
+* {
+    margin: 0;
+}
+.registro-container {
+    display: flex;
+    height: 100%;
+    flex-direction: column;
+}
+
+.registro-container > img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    position: absolute;
+    z-index: -1;
+}
+
+.background_container {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    background-color: rgba(245, 245, 245, 0.8);
+}
+
+#message {
+    position: absolute;
+    top: 0;
+    right: 0;
+    width: 20%;
+    height: 15%;
+    background-color: rgba(0, 131, 0, 0.5);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    color: white;
+    font-size: 1.5rem;
+    text-align: center;
+    margin: 10px;
+}
+
+#message.error {
+    background-color: rgba(255, 0, 0, 0.5);
+}
+
+.container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 35%;
+    height: auto;
+    margin: 10px 0;
+    background-color: whitesmoke;
+    border:  1px solid #e6e6e6;
+    border-radius: 5px;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+    transition: all 0.5s;
+    position: relative;
+}
+
+
+.container > h3 {
+    font-size: 30px;
+    font-weight: 600;
+    margin-top: 20px;
+}
+
+.container a {
+    color: var(--primary-blue);
+    text-decoration: none;
+    margin-bottom: 10px;
+}
+.container a:hover {
+    color: var(--dark-grey);
+    text-decoration: underline;
+}
+
+#idform {
+    display: flex;
+    flex-direction: column;
+    width: 80%;
+}
+
+.input-container {
+    display: flex;
+    flex-direction: column;
+    margin-bottom: 20px;
+}
+
+.input-container > input {
+    width: 100%;
+    height: 20px;
+    border: 1px solid #dbdbdb;
+    border-radius: 3px;
+    padding: 10px;
+    margin-bottom: 2px;
+    line-height: 30px;
+}
+
+.input-group-container {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+}
+
+.tipoUsuario {
+    display: flex;
+    flex-direction: row;
+    justify-content: left;
+    gap: 5px;
+    align-items: center;
+    margin: 20px 0;
+}
+.tipoUsuario > input {
+    margin-right: 5px;
+}
+
+#idform textarea {
+    resize: none;
+    line-height: 40px;
+}
+
+#idform > button {
+    width: 100%;
+    height: 40px;
+    margin-top: 10px;
+    margin-bottom: 10px;
+    border: 0;
+    border-radius: 3px;
+    background-color: var(--primary-blue);
+    color: white;
+    font-weight: 600;
+    cursor: pointer;
+}
+
+#idform > button:hover {
+    background-color: var(--secondary-blue);
+}
+
+@media screen and (max-height: 610px) {
+    .container {
+        width: 45%;
+        height: auto;
+        overflow-y: scroll;
+    }
+
+    .container > h3 {
+        font-size: 25px;
+    }
+
+    .input-container {
+        margin-bottom: 10px;
+    }
+}
+
